requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.1.0
colorama==0.4.5
colorlog==6.6.0
DateTime==4.4
frozenlist==1.3.0
hikari==2.0.0.dev109
hikari-lightbulb==2.2.2
idna==3.3
multidict==6.0.2
pytz==2022.1
yarl==1.7.2
zope.interface==5.4.0